tests/django_app/requirements.txt
huey==2.2.0
Django==3.1.12
redis==3.5.2
decorator==4.4.2
pytest-django==4.2.0

